(android) Add extensions methods for payment errors
Also moved existing extensions to separate packages.
