+++ phoenix-android/src/main/kotlin/fr/acinq/phoenix/android/intro/IntroView.kt
@@ -32,6 +32,7 @@ import fr.acinq.phoenix.android.*
 import fr.acinq.phoenix.android.R
 import fr.acinq.phoenix.android.components.*
 import fr.acinq.phoenix.android.utils.*
+import fr.acinq.phoenix.android.utils.extensions.findActivity
 import kotlinx.coroutines.*
