+++ phoenix-android/src/main/kotlin/fr/acinq/phoenix/android/payments/details/splash/SplashLightningOut.kt
@@ -41,7 +41,7 @@ import fr.acinq.phoenix.android.components.WebLink
 import fr.acinq.phoenix.android.components.contact.ContactOrOfferView
 import fr.acinq.phoenix.android.utils.Converter.toPrettyString
 import fr.acinq.phoenix.android.utils.MSatDisplayPolicy
-import fr.acinq.phoenix.android.utils.smartDescription
+import fr.acinq.phoenix.android.utils.extensions.smartDescription
 import fr.acinq.phoenix.data.LnurlPayMetadata
 import fr.acinq.phoenix.data.WalletPaymentId
 import fr.acinq.phoenix.data.WalletPaymentMetadata
