+++ phoenix-android/src/main/kotlin/fr/acinq/phoenix/android/security/EncryptedSeed.kt
@@ -16,14 +16,13 @@
 
 package fr.acinq.phoenix.android.security
 
-import fr.acinq.phoenix.android.utils.tryWith
+import fr.acinq.phoenix.android.utils.extensions.tryWith
 import fr.acinq.secp256k1.Hex
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import java.io.ByteArrayInputStream
 import java.io.ByteArrayOutputStream
 import java.security.GeneralSecurityException
-import javax.crypto.Cipher
 import javax.crypto.SecretKey
 
 sealed class EncryptedSeed {
