+++ phoenix-android/src/main/kotlin/fr/acinq/phoenix/android/utils/extensions/AndroidContextExtensions.kt
@@ -1,0 +1,39 @@
+/*
+ * Copyright 2024 ACINQ SAS
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package fr.acinq.phoenix.android.utils.extensions
+
+import android.content.Context
+import android.content.ContextWrapper
+import fr.acinq.phoenix.android.MainActivity
+
+fun Context.findActivitySafe(): MainActivity? {
+    var context = this
+    while (context is ContextWrapper) {
+        if (context is MainActivity) return context
+        context = context.baseContext
+    }
+    return null
+}
+
+fun Context.findActivity(): MainActivity {
+    var context = this
+    while (context is ContextWrapper) {
+        if (context is MainActivity) return context
+        context = context.baseContext
+    }
+    throw IllegalStateException("not in the context of the main Phoenix activity")
+}
