+++ phoenix-android/src/main/kotlin/fr/acinq/phoenix/android/utils/extensions/TechnicalExtensions.kt
@@ -1,0 +1,44 @@
+/*
+ * Copyright 2024 ACINQ SAS
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package fr.acinq.phoenix.android.utils.extensions
+
+import kotlin.contracts.ExperimentalContracts
+import kotlin.contracts.InvocationKind
+import kotlin.contracts.contract
+
+/**
+ * Utility method rebinding any exceptions thrown by a method into another exception, using the origin exception as the root cause.
+ * Helps with pattern matching.
+ */
+inline fun <T> tryWith(exception: Exception, action: () -> T): T = try {
+    action.invoke()
+} catch (t: Exception) {
+    exception.initCause(t)
+    throw exception
+}
+
+inline fun <T1 : Any, T2 : Any, R : Any> safeLet(p1: T1?, p2: T2?, block: (T1, T2) -> R?): R? {
+    return if (p1 != null && p2 != null) block(p1, p2) else null
+}
+
+@OptIn(ExperimentalContracts::class)
+inline fun <T, R> T.ifLet(block: (T) -> R): R {
+    contract {
+        callsInPlace(block, InvocationKind.EXACTLY_ONCE)
+    }
+    return block(this)
+}
